Added OCR ScreenSnip.ahk
OCR_ScreenSnip is an AHK v2 script that allows the user to take a screenshot using the Windows Snipping Tool. The OCR Library (from Descolada) then parses the screenshot and captures text found within. The script then adds it to the user's clipboard. 

Included in this commit is the OCR Library and examples from Descolda - https://github.com/Descolada/OCR/tree/main